requirements.txt
absl-py==0.4.1
appdirs==1.4.3
astor==0.7.1
bleach==1.5.0
captcha==0.2.4
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
colorama==0.4.1
cssselect==1.0.3
cycler==0.10.0
Cython==0.28.3
Flask==1.0.2
gast==0.2.0
gevent==1.3.6
greenlet==0.4.15
grpcio==1.14.2
h5py==2.7.1
html5lib==0.9999999
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Keras==2.0.8
kiwisolver==1.0.1
lxml==4.2.4
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
mpmath==1.0.0
numpy==1.14.5
pandas==0.22.0
Pillow==5.2.0
prettytable==0.7.2
protobuf==3.6.1
pycparser==2.19
pygame==1.9.3
PyMySQL==0.9.3
pynput==1.4
PyOpenGL==3.1.2
pyparsing==2.2.0
pyquery==1.4.0
python-dateutil==2.7.2
pytz==2018.4
PyYAML==3.12
requests==2.18.4
scikit-learn==0.19.2
scipy==1.0.1
selenium==3.141.0
six==1.11.0
sklearn==0.0
splinter==0.10.0
sympy==1.1.1
tensorboard==1.10.0
tensorflow-gpu==1.10.0
tensorflow-tensorboard==0.1.8
termcolor==1.1.0
tqdm==4.23.2
urllib3==1.22
Werkzeug==0.14.1
